+++ backend/src/main/java/com/unitcg/api/domain/produto/ProdutoRequestDTO.java
@@ -2,6 +2,6 @@ package com.unitcg.api.domain.produto;
 
 import org.springframework.web.multipart.MultipartFile;
 
-public record ProdutoRequestDTO(String name, String description, Double price, MultipartFile image) {
+public record ProdutoRequestDTO(String name, String description, int price, MultipartFile image) {
 
 }
